Remove 0.5 tests from modelgauge
