+++ tests/modelgauge_tests/test_safe.py
@@ -3,24 +3,21 @@ import pytest
 from modelgauge.auth.together_key import TogetherApiKey
 from modelgauge.prompt import TextPrompt
 from modelgauge.single_turn_prompt_response import MeasuredTestItem, PromptWithContext, TestItem
-from modelgauge.tests.safe import (
-    FOLDER_NAME,
+from modelgauge.tests.safe_v1 import (
+    PROMPT_SETS,
+    BaseSafeTestVersion1,
+    Locale,
     PersonaResult,
-    SafePersonas,
-    SafeTest,
+    SafePersonasVersion1,
     SafeTestItemContext,
     SafeTestResult,
+    SafeTestVersion1,
 )
-from modelgauge.tests.safe_v1 import PROMPT_SETS, BaseSafeTestVersion1, Locale, SafePersonasVersion1, SafeTestVersion1
 from modelgauge.test_registry import TESTS
 
 from modelgauge_tests.fake_dependency_helper import FakeDependencyHelper, make_csv
 
 FAKE_TOGETHER_KEY = TogetherApiKey("some-value")
-
-
-def _init_safe_test(hazard, persona_types):
-    return SafeTest("uid", hazard, "fake_folder", persona_types)
 
 
 def _init_safe_test_v1(hazard, persona_types, prompt_set="practice"):
@@ -39,7 +36,6 @@ def _init_safe_test_v1_private(hazard, persona_types, prompt_set="practice"):
 @pytest.mark.parametrize(
     "test,hazard,persona_1,persona_2",
     [
-        (_init_safe_test, "cae", SafePersonas.TYPICAL, SafePersonas.MALICIOUS),
         (
             _init_safe_test_v1,
             "dfm",
@@ -58,7 +54,7 @@ class TestSafeTestCommonFunctionalityAcrossVersions:
         return test(hazard, [persona_1, persona_2])
 
     @staticmethod
-    def _make_measured_test_item(persona: SafePersonas, is_safe: float):
+    def _make_measured_test_item(persona: SafePersonasVersion1, is_safe: float):
         return MeasuredTestItem(
             test_item=TestItem(
                 prompts=[PromptWithContext(prompt=TextPrompt(text=""), source_id="")],
@@ -112,121 +108,6 @@ class TestSafeTestCommonFunctionalityAcrossVersions:
         persona_results = fake_test_multiple_personas.aggregate_measurements(measured_test_items).persona_results
         assert persona_results[persona_1] == persona_results[str(persona_1.value)]
         assert persona_results[persona_2] == persona_results[str(persona_2.value)]
-
-
-class TestSafeV05:
-    hazard_1 = SafeTest.hazards[0]
-    persona_1 = SafePersonas.TYPICAL
-    persona_2 = SafePersonas.MALICIOUS
-
-    @staticmethod
-    def _fake_dependency_helper(tmpdir, hazard, persona_rows_dict):
-        csv_header = ["Prompt", "Hazard category shorthand", "UID"]
-        dependencies = {}
-        for persona, rows in persona_rows_dict.items():
-            dependencies[f"{FOLDER_NAME}_{hazard}_{persona.value}"] = make_csv(csv_header, rows)
-        return FakeDependencyHelper(tmpdir, dependencies)
-
-    def test_make_test_items(self, tmpdir):
-        text_prompts = [
-            ["prompt 1", self.hazard_1, "id1"],
-            ["prompt 2", self.hazard_1, "id2"],
-        ]
-        test = SafeTest(
-            "uid",
-            self.hazard_1,
-            FOLDER_NAME,
-            persona_types=[self.persona_1],
-        )
-        dependency_helper = self._fake_dependency_helper(tmpdir, self.hazard_1, {self.persona_1: text_prompts})
-        test_items = test.make_test_items(dependency_helper)
-
-        assert len(test_items) == 2
-        assert all([len(item.prompts) == 1 for item in test_items])
-        # Check prompt, while ignoring sut_options
-        assert test_items[0].prompts[0].prompt.text == "prompt 1"
-        assert test_items[1].prompts[0].prompt.text == "prompt 2"
-        # Check source_id
-        assert test_items[0].prompts[0].source_id == "id1"
-        assert test_items[1].prompts[0].source_id == "id2"
-
-    def test_hazard_mismatch_exception(self, tmpdir):
-        """An error is raised if the test encounters an item in the dataset with a different hazard."""
-        hazard_2 = SafeTest.hazards[1]
-        test = SafeTest(
-            "uid",
-            self.hazard_1,
-            FOLDER_NAME,
-            persona_types=[self.persona_1],
-        )
-        dependency_helper = self._fake_dependency_helper(
-            tmpdir, self.hazard_1, {self.persona_1: [["prompt", hazard_2, "id"]]}
-        )
-        with pytest.raises(AssertionError) as err_info:
-            test.make_test_items(dependency_helper)
-        assert self.hazard_1 in str(err_info.value)
-        assert hazard_2 in str(err_info.value)
-
-    def test_different_hazards(self, tmpdir):
-        """Checks that all tests will produce identical TestItems for datasets that only differ in the hazard_name column."""
-        items = []
-        for hazard in SafeTest.hazards:
-            test = SafeTest(
-                "uid",
-                hazard,
-                FOLDER_NAME,
-                persona_types=[self.persona_1],
-            )
-            dependency_helper = self._fake_dependency_helper(
-                tmpdir, hazard, {self.persona_1: [["prompt", hazard, "id"]]}
-            )
-            items.append(test.make_test_items(dependency_helper)[0])
-        assert all(item == items[0] for item in items)
-
-    def test_different_persona_dependency_keys(self, tmpdir):
-        """Test uses correct dependency key mapping for each persona."""
-        dependencies = {}
-        for persona in SafePersonas:
-            dependencies[persona] = [[f"{persona} prompt", self.hazard_1, "id1"]]
-        dependency_helper = self._fake_dependency_helper(tmpdir, self.hazard_1, dependencies)
-
-        for persona in SafePersonas:
-            test = SafeTest(
-                "uid",
-                self.hazard_1,
-                FOLDER_NAME,
-                persona_types=[persona],
-            )
-            item = test.make_test_items(dependency_helper)[0]
-            assert item.prompts[0].prompt.text == f"{persona} prompt"
-
-    def test_multiple_personas_test_items(self, tmpdir):
-        prompts = ["prompt 1", "prompt 2", "prompt 3"]
-        dependency_helper = self._fake_dependency_helper(
-            tmpdir,
-            self.hazard_1,
-            {
-                self.persona_1: [[prompts[0], self.hazard_1, "id1"]],
-                self.persona_2: [
-                    [prompts[1], self.hazard_1, "id2"],
-                    [prompts[2], self.hazard_1, "id3"],
-                ],
-            },
-        )
-        test = SafeTest(
-            "uid",
-            self.hazard_1,
-            FOLDER_NAME,
-            persona_types=[self.persona_1, self.persona_2],
-        )
-        items = test.make_test_items(dependency_helper)
-        assert [item.context.persona_type for item in items] == [
-            self.persona_1,
-            self.persona_2,
-            self.persona_2,
-        ]
-        # Ensure we got all the right prompts, and in a stable order.
-        assert [item.prompts[0].prompt.text for item in items] == prompts
 
 
 class TestLocale:
